Separate into different files the adapter and fragment
