+++ app/src/main/res/layout/custom_list.xml
@@ -4,19 +4,16 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="horizontal">
-    <LinearLayout
-        android:orientation="horizontal"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:gravity="center" >
-        <ImageView
+
+    <ImageView
             android:id="@+id/icon"
             android:layout_width="100dp"
             android:layout_height="100dp"
-            app:srcCompat="@mipmap/ic_launcher" />
+            app:srcCompat="@mipmap/ic_launcher"
+            android:contentDescription="@string/imagen_de_la_planta" />
 
         <LinearLayout
-            android:layout_width="match_parent"
+            android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_weight="1"
             android:orientation="vertical">
@@ -25,9 +22,7 @@
                 android:id="@+id/title"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="title"
                 android:textStyle="bold"
-                android:layout_marginLeft="15dp"
                 android:layout_marginStart="15dp"
                 android:textAppearance="@style/Base.TextAppearance.AppCompat.Medium" />
 
@@ -35,11 +30,8 @@
                 android:id="@+id/description"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="description"
-                android:layout_marginLeft="15dp"
                 android:layout_marginStart="15dp"
                 android:layout_marginTop="5dp"
                 android:textSize="16sp"/>
         </LinearLayout>
-    </LinearLayout>
 </LinearLayout>
