+++ app/src/main/res/values/strings.xml
@@ -7,4 +7,6 @@
     <string name="reminders">Reminders</string>
     <string name="tomar_foto">Take picture</string>
     <string name="siguiente">Next</string>
+    <string name="Probabibilidad">Probability: %1$s</string>
+    <string name="imagen_de_la_planta">Imagen de la planta</string>
 </resources>
